+++ packages/select/src/components/index.ts
@@ -5,7 +5,6 @@
  */
 
 export * from "./omnibar/omnibar";
-export * from "./query-list/itemRenderer";
 export * from "./query-list/queryList";
 export * from "./select/multiSelect";
 export * from "./select/select";
