+++ packages/select/src/index.ts
@@ -5,5 +5,6 @@
  */
 
 export * from "./common/classes";
+export * from "./common/itemRenderer";
 export * from "./common/predicate";
 export * from "./components";
